@@ -1,3 +1,3 @@
 module goposts
 
-go 1.21.7
+go 1.22.0
